move redux provider mock to test utils folder
